ypdate"21/04"

+++ src/main/java/com/example/demo/repository/MedecinRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.example.demo.models.Medecin;
 
-public interface MedecinRepository extends JpaRepository<Medecin, Integer> {
+public interface MedecinRepository extends JpaRepository<Medecin, Long> {
 	
 	@Query("select m from Medecin m where m.email =:email and m.password =:password")
 	Medecin findByEmailAndPassword (@Param("email") String email,@Param("password") String password);
