+++ src/main/java/com/example/demo/services/PatientServiceInterface.java
@@ -8,8 +8,8 @@ public interface PatientServiceInterface {
 
 	Patient UpdatePatient(Patient p);
 	void deletePatient(Patient p);
-	void deletePatientById(int id);
-	Patient getPatient(int id);
+	void deletePatientById(Long id);
+	Patient getPatient(Long id);
 	List<Patient> getAllPatients();
     Patient savePatient(Patient p);
 	List<Patient> afficherPatient();
